% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criterias.R
\name{AIC}
\alias{AIC}
\alias{BIC}
\alias{BIC_Q}
\alias{BIC_adj}
\alias{CAIC}
\alias{KIC}
\alias{AIC_4}
\alias{HQIC}
\alias{FIC}
\alias{ICOMP_IFIM_CF}
\alias{ICOMP_IFIM_C1}
\alias{ICOMP_IFIM_C1F}
\alias{ICOMP_IFIM_C1R}
\alias{ICOMP_PEU_CF}
\alias{ICOMP_PEU_C1}
\alias{ICOMP_PEU_C1F}
\alias{ICOMP_PEU_C1R}
\alias{ICOMP_PEU_LN_CF}
\alias{ICOMP_PEU_LN_C1}
\alias{ICOMP_PEU_LN_C1F}
\alias{ICOMP_PEU_LN_C1R}
\alias{CAIC_F}
\alias{JIC}
\alias{IBIC}
\alias{SPBIC}
\alias{HBIC}
\alias{GCV}
\title{Information Criteria}
\usage{
AIC(model)

BIC(model)

BIC_Q(model, q = 0.5)

BIC_adj(model, q = 0.5)

CAIC(model)

KIC(model)

AIC_4(model)

HQIC(model)

FIC(model)

ICOMP_IFIM_CF(model)

ICOMP_IFIM_C1(model)

ICOMP_IFIM_C1F(model)

ICOMP_IFIM_C1R(model)

ICOMP_PEU_CF(model)

ICOMP_PEU_C1(model)

ICOMP_PEU_C1F(model)

ICOMP_PEU_C1R(model)

ICOMP_PEU_LN_CF(model)

ICOMP_PEU_LN_C1(model)

ICOMP_PEU_LN_C1F(model)

ICOMP_PEU_LN_C1R(model)

CAIC_F(model)

JIC(model)

IBIC(model)

SPBIC(model)

HBIC(model)

GCV(model)
}
\arguments{
\item{model}{a "lm" or "glm" object}
}
\value{
Information Criteria value
}
\description{
Various information criterias for "lm" and "glm" objects.
}
\details{
\code{AIC} is calculated using

-2LL + 2k

\code{BIC}
asdasdsad
}
