% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_fisher.R
\name{reverse_fisher}
\alias{reverse_fisher}
\title{Reverse Fisher Matrix}
\usage{
reverse_fisher(model)
}
\arguments{
\item{model}{a "lm" or "glm" object}
}
\value{
a booster object with below components.
 \item{n_train}{Number of cases in the input dataset.}
}
\description{
This function allows you to calculate Fisher Information Matrix using "lm" and "glm" objects.
}
\details{
Calculates Fisher Information Matrix using "lm" and "glm" objects. It uses
}
\keyword{internal}
